mcp-server: one benchmark assessor — absence is never credited
The MCP hackmyagent_benchmark tool's interim assessor kept the legacy
reading #458 names: a control none of whose checkIds produced any
record counted as PASSED (.get(id) !== false over an empty map), so an
assessment fed nothing read 100% compliance with every control passing.

generateBenchmarkReport — the CLI benchmark's function, extracted to
src/benchmarks/benchmark-report.ts because importing cli.ts is
side-effectful — is now the one assessor for both surfaces:
a control with no record is unverified, never credited; a level where
nothing measured reports "not measured" / Not Assessed (step 0's
contract), not a fabricated 0%; step 3's not-applicable and
measured-wins semantics apply identically. The interim bucket and its
pinned tests are replaced by ruled-semantics pins, as the interim's
own doc comment scheduled. HTML gives not-applicable controls their
own status class.

Closes #458.
